requirements.txt
black==21.11b1
python-daemon==2.3.0
PyYAML==6.0
